Added current playing song to bar

+++ install.sh
@@ -2,6 +2,17 @@
 
 DOTDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
+
+# Confirm actions
+read -p 'This will erase all current dotfiles and the contents of ~/bin. Are you sure you want to continue? (Y/N)' answer
+accept=0
+case "${answer}" in [yY]|[yY][eE][sS])
+    accept=1
+esac
+if [[ $accept = 0 ]]; then
+    exit
+fi
+
 # Reset Current Files
 rm -f ~/.config/i3/config
 rm -f ~/.config/i3/i3blocks.conf
@@ -10,6 +21,7 @@ rm -f ~/.vimrc
 rm -f ~/.xinitrc
 rm -f ~/.Xresources
 rm -f ~/.zshrc
+rm -rf ~/bin
 echo "Reset current files"
 
 # Link New Files
@@ -20,6 +32,7 @@ ln -s ${DOTDIR}/vimrc ~/.vimrc
 ln -s ${DOTDIR}/xinitrc ~/.xinitrc
 ln -s ${DOTDIR}/Xresources ~/.Xresources
 ln -s ${DOTDIR}/zshrc ~/.zshrc
+ln -s ${DOTDIR}/bin ~/
 echo "Linked new files"
 
 echo "Please read readme for further installation!"
